edit some

+++ client/src/pages/Chat/Chat.jsx
@@ -1,70 +1,79 @@
-import { UilSetting } from '@iconscout/react-unicons'
+// Chat.jsx
 import React, { useEffect, useRef, useState } from 'react'
-import { useSelector } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
+import io from 'socket.io-client'
 import { Link } from 'react-router-dom'
-import { io } from 'socket.io-client'
-import { createChat, userChats, deleteChatRequest } from '../../api/ChatRequest'
+import {
+  clearMessages,
+  connectSocket,
+  disconnectSocket,
+  sendMessage,
+  setChats,
+  setCurrentChat,
+} from '../../actions/socketAction'
+import {
+  userChats,
+  createChat as createChatRequest,
+  deleteChatRequest,
+} from '../../api/ChatRequest'
 import ChatBox from '../../components/ChatBox/ChatBox'
 import Conversation from '../../components/Conversation/Conversation'
-import Comment from '../../img/comment.png'
 import Home from '../../img/home.png'
-import Noti from '../../img/noti.png'
 import LogoSearch from './../../components/LogoSearch/LogoSearch'
 import './Chat.scss'
 import SearchForm from '../../components/SearchForm/SearchForm'
 import ConversationInSearch from '../../components/ConversationInSearch/ConversationInSearch'
 import { searchUsers } from '../../api/UserRequest'
 import LeftSideBar from '../../components/LeftSideBar/LeftSideBar'
+import { UilSetting } from '@iconscout/react-unicons'
+
 const Chat = () => {
-  const { user } = useSelector((state) => state.authReducer.authData)
-  const [chats, setChats] = useState([])
-  const [currentChat, setCurrentChat] = useState(null)
-  const [onlineUsers, setOnlineUsers] = useState([])
-  const [sendMessage, setSendMessage] = useState(null)
-  const [receiveMessage, setReceiveMessage] = useState(null)
+  const dispatch = useDispatch()
+  const authData = useSelector((state) => state.authReducer.authData) || {
+    user: null,
+  }
+  const { user } = authData
+  const { chats, currentChat, onlineUsers, receiveMessage } = useSelector(
+    (state) => state.socketReducer
+  )
   const [showChat, setShowChat] = useState(true)
   const [userSearchs, setUserSearchs] = useState([])
   const [querySearch, setQuerySearch] = useState('')
-  const socket = useRef()
+  const socket = useRef(null)
 
   useEffect(() => {
-    socket.current = io('ws://localhost:8800')
-    socket.current.emit('new-user-add', user._id)
-    socket.current.on('get-users', (users) => {
-      setOnlineUsers(users)
-    })
-  }, [user])
-  // send message to socket server
-  useEffect(() => {
-    if (sendMessage !== null) {
-      socket.current.emit('send-message', sendMessage)
+    if (user) {
+      socket.current = io('ws://localhost:8800')
+      dispatch(connectSocket(user._id))
     }
-  }, [sendMessage])
-
-  // receive message from socket server
-  useEffect(() => {
-    socket.current.on('receive-message', (data) => {
-      setReceiveMessage(data)
-    })
-  }, [])
+    return () => {
+      if (socket.current) {
+        socket.current.disconnect()
+      }
+      dispatch(disconnectSocket())
+    }
+  }, [user, dispatch])
 
   useEffect(() => {
     const getChats = async () => {
       try {
         const { data } = await userChats(user._id)
-        setChats(data)
+        dispatch(setChats(data))
       } catch (error) {
         console.log(error)
       }
     }
     getChats()
-  }, [user])
+  }, [user, dispatch])
 
   useEffect(() => {
     const getUsers = async () => {
       try {
         const { data } = await searchUsers(querySearch)
-        setUserSearchs(data)
+        const userSearchs = data.filter(
+          (userSearch) => userSearch._id !== user._id
+        )
+        setUserSearchs(userSearchs)
       } catch (error) {
         console.log(error)
       }
@@ -84,19 +93,35 @@ const Chat = () => {
 
   const handleUserClick = async (receiverId) => {
     try {
-      const newChat = await createChat(user._id, receiverId)
-      setChats((prevChats) => [...prevChats, newChat])
-      setCurrentChat(newChat)
+      const { data } = await createChatRequest(user._id, receiverId)
+      if (data.isExist) {
+        // Handle existing chat
+        dispatch(setCurrentChat(data.existingChat))
+      } else {
+        // Handle new chat
+        dispatch(setCurrentChat(data))
+        socket.current.emit('create-chat', {
+          senderId: user._id,
+          receiverId,
+          chat: data,
+        })
+      }
     } catch (error) {
       console.error('Failed to create chat', error)
     }
   }
 
-  const deleteChat = async (chatId) => {
+  const handleDeleteChat = async (chatId) => {
     try {
-      await deleteChatRequest(chatId)
-      setChats((prevChats) => prevChats.filter((chat) => chat._id !== chatId))
-      setCurrentChat(null)
+      const chatToDelete = chats.find((chat) => chat._id === chatId)
+      if (chatToDelete) {
+        await deleteChatRequest(chatId)
+        socket.current.emit('delete-chat', {
+          chatData: chatToDelete,
+        })
+        dispatch(setCurrentChat(null))
+        dispatch(clearMessages())
+      }
     } catch (error) {
       console.error('Failed to delete chat', error)
     }
@@ -104,10 +129,11 @@ const Chat = () => {
 
   return (
     <div className="Chat">
-      <div className="left-sidebar">
-        <LeftSideBar />
+      <div className="left-sidebar responsive-sidebar">
+        <div className="hidden md:block">
+          <LeftSideBar />
+        </div>
       </div>
-      {/* Left side */}
       <div className="Left-side-chat">
         <LogoSearch />
         <div className="Chat-container">
@@ -118,73 +144,70 @@ const Chat = () => {
           <div className="Chat-list hover-scrollbar">
             {showChat
               ? chats.map((chat) => (
-                <div
-                  key={chat._id}
-                  role="button"
-                  tabIndex={0}
-                  onClick={() => setCurrentChat(chat)}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter') {
-                      setCurrentChat(chat)
-                    }
-                  }}
-                  className="Conversation"
-                >
-                  <Conversation
-                    data={chat}
-                    currentUserId={user._id}
-                    online={checkOnlineStatus(chat)}
-                  />
-                  <button
-                    className="delete-button"
-                    onClick={() => {
-                      deleteChat(chat._id)
+                  <div
+                    key={chat?._id}
+                    role="button"
+                    tabIndex={0}
+                    onClick={() => dispatch(setCurrentChat(chat))}
+                    onKeyDown={(e) => {
+                      if (e.key === 'Enter') {
+                        dispatch(setCurrentChat(chat))
+                      }
+                    }}
+                    className="Conversation"
+                  >
+                    <Conversation
+                      data={chat}
+                      currentUserId={user._id}
+                      online={checkOnlineStatus(chat)}
+                    />
+                    <button
+                      className="delete-button"
+                      onClick={(e) => {
+                        e.stopPropagation()
+                        handleDeleteChat(chat._id)
+                      }}
+                    >
+                      X
+                    </button>
+                  </div>
+                ))
+              : userSearchs.map((user) => (
+                  <div
+                    key={user._id}
+                    role="button"
+                    tabIndex={0}
+                    onClick={() => handleUserClick(user._id)}
+                    onKeyDown={(e) => {
+                      if (e.key === 'Enter') {
+                        handleUserClick(user._id)
+                      }
                     }}
                   >
-                    X
-                  </button>
-                </div>
-              ))
-              : userSearchs.map((user) => (
-                <div
-                  key={user._id}
-                  role="button"
-                  tabIndex={0}
-                  onClick={() => handleUserClick(user._id)}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter') {
-                      handleUserClick(user._id)
-                    }
-                  }}
-                >
-                  <ConversationInSearch
-                    data={user}
-                    setShowChat={setShowChat}
-                  />
-                </div>
-              ))}
+                    <ConversationInSearch
+                      data={user}
+                      setShowChat={setShowChat}
+                    />
+                  </div>
+                ))}
           </div>
         </div>
       </div>
-      {/* Right side */}
       <div className="Right-side-chat">
-        <div style={{ width: '20rem', alignSelf: 'flex-end' }}>
+        <div style={{ width: '8rem', alignSelf: 'flex-end' }}>
           <div className="navIcons">
             <Link to="/dashhome">
               <img src={Home} alt="" />
             </Link>
-            <UilSetting />
-            <img src={Noti} alt="" />
-            <Link to="/chat">
-              <img src={Comment} alt="" />
+            <Link to={`/profile/${user._id}`}>
+              <UilSetting />
             </Link>
           </div>
         </div>
-        {/* Chat body */}
         <ChatBox
           chat={currentChat}
           currentUser={user._id}
-          setSendMessage={setSendMessage}
+          setSendMessage={(message) => dispatch(sendMessage(message))}
           receiveMessage={receiveMessage}
         />
       </div>
